Sequelize models created
