+++ package.json
@@ -4,7 +4,7 @@
     "description": "Task Management",
     "main": "index.js",
     "scripts": {
-        "dev": "nodemon --watch server --exec \"node --inspect server/index.js\"",
+        "dev": "nodemon --watch server --watch .env --exec \"node --inspect server/index.js\"",
         "prod": "node server/index.js",
         "lint": "./node_modules/eslint-watch/bin/esw --color",
         "lint:watch": "npm run lint --watch"
@@ -26,10 +26,15 @@
         "mongoose": "^5.9.9",
         "morgan": "^1.10.0",
         "multer": "^1.4.2",
+        "mysql2": "^2.1.0",
         "nodemailer": "^6.4.6",
         "nodemailer-sendgrid-transport": "^0.2.0",
         "npm": "^6.14.5",
         "npm-run-all": "^4.1.5",
+        "pg": "^8.2.1",
+        "pg-hstore": "^2.3.3",
+        "sequelize": "^5.21.11",
+        "umzug": "^2.3.0",
         "winston": "^3.2.1",
         "winston-daily-rotate-file": "^4.4.2"
     },
