+++ server/bootstrap/Express.js
@@ -37,7 +37,7 @@ const init = async (moduleObj) => {
         app.use("/api", route.router);
     }
 
-    const PORT = process.env.PORT || 8080;
+    const PORT = process.env.PORT;
     app.listen(PORT, () => {
         console.log(`App listening to ${PORT}....`);
         console.log("Press Ctrl+C to quit.");
