+++ server/config/index.js
@@ -1,0 +1,5 @@
+const db = require("./db");
+
+module.exports = {
+    db
+};
