+++ server/database/migrate.js
@@ -1,0 +1,3 @@
+const sequelize = require("./sequelize");
+
+sequelize.sync({ alter: true });
